refactor(map): extract useMapAiProposals FSM hook (5/7)
Step 5 of the PulseMapView strangler decomposition. Pure move — no
behaviour change. The state machine, the 300ms debounce + AbortController
contract, the Directions service call, the reorder lifecycle, and the
Open-in-Maps URL builder are all preserved byte-equivalent.

hooks/useMapAiProposals.ts
  - Owns aiState (idle / fetching / ready / none / paused / reordering),
    acceptedRoute, acceptingRoute, and isReordering (derived).
  - The proposal effect debounces 300ms, aborts on dep change, and
    silently collapses null proposals to 'paused' (when getAiPausedUntil
    returns a timestamp) or 'none' (everything else). Skip-guards for
    accepted-route and reordering states preserve the prior "don't wipe
    the underway / draft state on input change" behaviour.
  - handleAcceptRoute is dual-source — orderedIds come from either the
    fresh ready-route proposal or the reorder draft. Runs DirectionsService
    once, snapshots path + duration + arrivesAt into AcceptedRoute.
  - handleStartReorder / Change / Cancel implement the FSM transitions
    ready ↔ reordering without losing the base proposal (Cancel reverts).
  - handleOpenInSystemMaps builds a google.com/maps/dir URL targeting
    the first stop.
  - reorderableStops memo strips lat/lng off the unified stops list for
    the AiStrip's drag UI.

PulseMapView.tsx
  - Three useState declarations (aiState, acceptedRoute, acceptingRoute)
    and ~210 lines of effect + handlers collapse to one hook call.
  - Drops getAiPausedUntil, proposeAtlasInsight, proposeRoute,
    proposeWeekPlan, RouteProposal imports — all only used inside the
    extracted hook. DAY_MS drops too (the staleDays calculation that
    used it moved with the Atlas snapshot logic).
  - End-of-file e2e re-exports keep the harness import surface:
    `export type { AiState, AcceptedRoute } from './sub/aiTypes'` is the
    direct re-export now (no host import roundtrip required).
  - PulseMapView still owns allStops (consumed by the hook AND the
    marker render path), markerKey / meetingKey helpers (shared between
    allStops and marker rendering), and the visibleMarkers memo.
