+++ src/components/map/PulseMapView.tsx
@@ -5,13 +5,6 @@ import { AppView, CalendarEvent, Contact } from '../../types';
 import { ContactCircle } from '../../types/contactCircleTypes';
 import { GOOGLE_MAPS_LIBRARIES, convexHull, getMapOptions, computeBounds } from '../../services/mapService';
 import { UserLocation } from '../../services/locationService';
-import {
-  getAiPausedUntil,
-  proposeAtlasInsight,
-  proposeRoute,
-  proposeWeekPlan,
-  RouteProposal,
-} from '../../services/mapAIService';
 import MapFilterBar, { MapFilter } from './MapFilterBar';
 import MapContactMarker from './contacts/MapContactMarker';
 import MapRadiusRings from './contacts/MapRadiusRings';
@@ -21,7 +14,6 @@ import MapMeetingMarker from './contacts/MapMeetingMarker';
 import LocationEditModal from './contacts/LocationEditModal';
 import ImAtFAB from './sub/ImAtFAB';
 import {
-  DAY_MS,
   DEFAULT_CENTER,
   DEFAULT_ZOOM,
   LENS_OPTIONS,
@@ -30,7 +22,6 @@ import {
   type MapLens,
   type MapViewMode,
 } from './sub/mapLens';
-import type { AcceptedRoute, AiState } from './sub/aiTypes';
 import { AiStrip } from './sub/AiStrip';
 import { LensEmptyState } from './sub/LensEmptyState';
 import { LiveBroadcastSheet } from './sub/LiveBroadcastSheet';
@@ -41,6 +32,7 @@ import { useVisitedStops } from './hooks/useVisitedStops';
 import { useUserPosition } from './hooks/useUserPosition';
 import { useContactGeocoding } from './hooks/useContactGeocoding';
 import { useLivePresence } from './hooks/useLivePresence';
+import { useMapAiProposals } from './hooks/useMapAiProposals';
 
 interface PulseMapViewProps {
   contacts: Contact[];
@@ -98,15 +90,6 @@ const PulseMapView: React.FC<PulseMapViewProps> = ({
     setViewMode(next);
     try { localStorage.setItem(MAP_VIEW_LS_KEY, next); } catch { /* ignore */ }
   }, []);
-
-  // AI strip state machine (idle → fetching → ready/none). See the hoisted
-  // types above the component for the proposal shape.
-  const [aiState, setAiState] = useState<AiState>({ status: 'idle' });
-
-  // Populated when the user clicks Accept on a TODAY route proposal. Drives
-  // the polyline overlay, the marker renumbering, and the "Underway" strip.
-  const [acceptedRoute, setAcceptedRoute] = useState<AcceptedRoute | null>(null);
-  const [acceptingRoute, setAcceptingRoute] = useState(false);
 
   const { isLoaded, loadError } = useJsApiLoader({
     id: 'pulse-google-maps',
@@ -264,217 +247,30 @@ const PulseMapView: React.FC<PulseMapViewProps> = ({
     [visibleMarkers, meetingMarkers, geoSignals.todayEvents, geoSignals.weekEvents],
   );
 
-  // ─── AI proposal effect — paint-first, fetch in the background, collapse
-  // silently on failure. Aborts on lens change or new marker set. Skips when
-  // a route is already accepted (the strip then renders the Underway state)
-  // or when the user is actively reordering a proposal (would wipe their
-  // draft on every marker filter change). ─────────────────────────────────────
-  const isReordering = aiState.status === 'reordering';
-  useEffect(() => {
-    if (!isLoaded) return;
-    if (acceptedRoute) return; // Underway — don't re-propose.
-    if (isReordering) return;  // User is mid-drag — preserve the working order.
-    if (allStops.length === 0) {
-      setAiState({ status: 'idle' });
-      return;
-    }
-
-    const controller = new AbortController();
-    setAiState({ status: 'fetching' });
-
-    const debounceTimer = setTimeout(async () => {
-      if (controller.signal.aborted) return;
-
-      // Use the unified stops list — contact markers PLUS meeting markers
-      // resolved from calendar events with a location. The AI gets both,
-      // letting it weave meetings into the route order (e.g. "1pm meeting at
-      // X first, then the cluster of contacts nearby, then home stop").
-      const stops = allStops.map(s => ({
-        id: s.id,
-        label: s.label,
-        lat: s.lat,
-        lng: s.lng,
-      }));
-
-      // Helper — null proposal could mean "no insight" OR "circuit open
-      // because we hit the workspace cap." Surface paused explicitly so the
-      // user knows AI is on a break, not broken.
-      const settleNullProposal = () => {
-        const pausedUntil = getAiPausedUntil();
-        if (pausedUntil != null) setAiState({ status: 'paused', until: pausedUntil });
-        else setAiState({ status: 'none' });
-      };
-
-      try {
-        if (lens === 'today') {
-          if (stops.length < 2) {
-            setAiState({ status: 'none' });
-            return;
-          }
-          const proposal = await proposeRoute({
-            stops,
-            origin: userPosition,
-            signal: controller.signal,
-            visitedIds: Array.from(visitedStopIds),
-          });
-          if (controller.signal.aborted) return;
-          if (!proposal) { settleNullProposal(); return; }
-          setAiState({ status: 'ready', data: { kind: 'route', proposal } });
-          return;
-        }
-        if (lens === 'week') {
-          const proposal = await proposeWeekPlan({ stops, signal: controller.signal });
-          if (controller.signal.aborted) return;
-          if (!proposal) { settleNullProposal(); return; }
-          setAiState({ status: 'ready', data: { kind: 'plan', proposal } });
-          return;
-        }
-        // Atlas — insight from the network at large.
-        const snapshots = visibleMarkers.map(m => {
-          const staleMs = m.contact.lastSeen ? Date.now() - m.contact.lastSeen.getTime() : null;
-          return {
-            id: m.contact.id,
-            name: m.contact.name,
-            lat: m.lat,
-            lng: m.lng,
-            staleDays: staleMs != null ? Math.floor(staleMs / DAY_MS) : null,
-            circles: circles
-              .filter(c => c.memberContactIds.includes(m.contact.id))
-              .map(c => c.name),
-          };
-        });
-        const proposal = await proposeAtlasInsight({ contacts: snapshots, signal: controller.signal });
-        if (controller.signal.aborted) return;
-        if (!proposal) { settleNullProposal(); return; }
-        setAiState({ status: 'ready', data: { kind: 'insight', proposal } });
-      } catch {
-        if (!controller.signal.aborted) settleNullProposal();
-      }
-    }, 300);
-
-    return () => {
-      clearTimeout(debounceTimer);
-      controller.abort();
-    };
-    // allStops includes meeting markers — they need to participate in the AI
-    // proposal alongside contact markers. visitedStopIds suppresses already-
-    // arrived stops from the prompt so the AI doesn't re-route to them.
-  }, [isLoaded, lens, allStops, userPosition, circles, acceptedRoute, isReordering, visitedStopIds]);
-
-  // ─── Route accept — calls the Google Directions JS service for the polyline
-  // and the leg-duration math, then flips the strip into the Underway state.
-  // Accepts from either a fresh ready-route proposal OR a user-reordered draft
-  // (same DirectionsService call, just a different orderedIds source). ─────────
-  const handleAcceptRoute = useCallback(async () => {
-    if (acceptingRoute) return;
-    let orderedIds: string[] | null = null;
-    if (aiState.status === 'reordering') {
-      orderedIds = aiState.orderedIds;
-    } else if (aiState.status === 'ready' && aiState.data.kind === 'route') {
-      orderedIds = aiState.data.proposal.orderedIds;
-    } else {
-      return;
-    }
-    if (orderedIds.length < 2) return;
-
-    // Resolve every ordered id from the unified allStops list — handles both
-    // contact markers and meeting markers without forking the lookup.
-    // Defensive filter for ids that vanished between proposal and accept
-    // (filter change, reorder against a stale set, etc.).
-    const orderedResolved = orderedIds
-      .map(id => allStops.find(s => s.id === id))
-      .filter((s): s is typeof allStops[number] => s != null);
-    if (orderedResolved.length < 2) return;
-
-    const origin = userPosition ?? { lat: orderedResolved[0].lat, lng: orderedResolved[0].lng };
-    const destination = {
-      lat: orderedResolved[orderedResolved.length - 1].lat,
-      lng: orderedResolved[orderedResolved.length - 1].lng,
-    };
-    const waypoints = orderedResolved.slice(0, -1).map(s => ({
-      location: { lat: s.lat, lng: s.lng },
-      stopover: true,
-    }));
-
-    setAcceptingRoute(true);
-    try {
-      const ds = new google.maps.DirectionsService();
-      const result = await ds.route({
-        origin,
-        destination,
-        waypoints,
-        travelMode: google.maps.TravelMode.DRIVING,
-        optimizeWaypoints: false, // AI already proposed the order
-      });
-      const route = result.routes[0];
-      if (!route) {
-        setAcceptingRoute(false);
-        return;
-      }
-      const path = route.overview_path.map(p => ({ lat: p.lat(), lng: p.lng() }));
-      const totalSec = route.legs.reduce((acc, leg) => acc + (leg.duration?.value ?? 0), 0);
-      const durationMin = Math.round(totalSec / 60);
-      setAcceptedRoute({
-        orderedMarkerKeys: orderedResolved.map(s => s.id),
-        path,
-        durationMin,
-        arrivesAt: new Date(Date.now() + totalSec * 1000),
-      });
-    } catch {
-      // Directions failed — leave the strip as-is so the user can retry.
-    } finally {
-      setAcceptingRoute(false);
-    }
-  }, [aiState, acceptingRoute, allStops, userPosition]);
-
-  const handleDismissRoute = useCallback(() => {
-    setAcceptedRoute(null);
-  }, []);
-
-  // ─── Reorder lifecycle. Enter from a ready route proposal, mutate the draft
-  // order via drag, then either accept the new order (Accept reuses
-  // handleAcceptRoute which now consumes the reorder draft) or revert. ────────
-  const handleStartReorder = useCallback(() => {
-    if (aiState.status !== 'ready' || aiState.data.kind !== 'route') return;
-    setAiState({
-      status: 'reordering',
-      orderedIds: aiState.data.proposal.orderedIds.slice(),
-      baseProposal: aiState.data.proposal,
-    });
-  }, [aiState]);
-
-  const handleReorderChange = useCallback((nextOrder: string[]) => {
-    setAiState(prev =>
-      prev.status === 'reordering' ? { ...prev, orderedIds: nextOrder } : prev,
-    );
-  }, []);
-
-  const handleReorderCancel = useCallback(() => {
-    setAiState(prev =>
-      prev.status === 'reordering'
-        ? { status: 'ready', data: { kind: 'route', proposal: prev.baseProposal } }
-        : prev,
-    );
-  }, []);
-
-  // Stops the AiStrip's reorder list renders. Uses the unified allStops so
-  // meetings show up in the drag list alongside contact stops. Mapped to
-  // {id, label} only — the strip doesn't need lat/lng.
-  const reorderableStops = useMemo(
-    () => allStops.map(s => ({ id: s.id, label: s.label })),
-    [allStops],
-  );
-
-  // Open the first leg in the OS maps app — closest current platform-native
-  // affordance to "tap to start" without bundling Apple/Google Maps SDKs.
-  const handleOpenInSystemMaps = useCallback(() => {
-    if (!acceptedRoute) return;
-    const firstStop = allStops.find(s => s.id === acceptedRoute.orderedMarkerKeys[0]);
-    if (!firstStop) return;
-    const destStr = `${firstStop.lat},${firstStop.lng}`;
-    const url = `https://www.google.com/maps/dir/?api=1&destination=${encodeURIComponent(destStr)}&travelmode=driving`;
-    window.open(url, '_blank', 'noopener');
-  }, [acceptedRoute, allStops]);
+  // AI proposal FSM + accept/reorder/dismiss/open-in-maps handlers all live
+  // in useMapAiProposals. The hook owns aiState, acceptedRoute, and
+  // acceptingRoute; the strip prop wiring below stays prop-for-prop identical
+  // to the prior inline implementation.
+  const {
+    aiState,
+    acceptedRoute,
+    acceptingRoute,
+    reorderableStops,
+    handleAcceptRoute,
+    handleDismissRoute,
+    handleStartReorder,
+    handleReorderChange,
+    handleReorderCancel,
+    handleOpenInSystemMaps,
+  } = useMapAiProposals({
+    isLoaded,
+    lens,
+    allStops,
+    userPosition,
+    circles,
+    visitedStopIds,
+    visibleMarkers,
+  });
 
   const selectedContact = localContacts.find(c => c.id === selectedContactId) ?? null;
 
@@ -1066,4 +862,4 @@ export default PulseMapView;
 // These components are otherwise internal — exporting them avoids the harness
 // having to recreate dialog/reorder UI to test focus management and a11y.
 export { AiStrip, LiveBroadcastSheet };
-export type { AiState, AcceptedRoute };
+export type { AiState, AcceptedRoute } from './sub/aiTypes';
